% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analysis_fixed_effects}
\alias{analysis_fixed_effects}
\title{Calculates covariance/variance matrix for DBC and alpha lattice design in
single environment}
\usage{
analysis_fixed_effects(df, design, multi_env)
}
\arguments{
\item{df}{data.frame}

\item{design}{character defining "DBC" do "lattice"}

\item{multi_env}{logica TRUE/FALSE indicating multienvironment analysis}
}
\description{
Calculates covariance/variance matrix for DBC and alpha lattice design in
single environment
}
