% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{elston_index}
\alias{elston_index}
\title{Elston selection index}
\usage{
elston_index(df, k = NULL, increasing = NULL)
}
\arguments{
\item{df}{data.frame with first column called 'gen' with genotypes identification and follow columns with adjusted mean for each phenotype.}

\item{k}{vector with minimum or maximum value for each phenotype evaluated. If NULL the minimum/maximum value is considered.}

\item{increasing}{vector with column names of phenotypes that low value is advantageous}
}
\description{
Elston selection index
}
