% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mulamba_index}
\alias{mulamba_index}
\title{Mulamba and Mock (1978) selection index}
\usage{
mulamba_index(df, increasing = NULL, weights = NULL)
}
\arguments{
\item{df}{data.frame with first column called 'gen' with genotypes identification and follow columns with adjusted mean for each phenotype.}

\item{increasing}{vector with column names of phenotypes that low value is advantageous}

\item{weights}{vector of economic weight for each phenotype. If df has 8 columns ('gen' + 7 phenotypes), this vector should have length 7.}
}
\description{
Mulamba and Mock (1978) selection index
}
