% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assumptionsTest.R
\name{outlier}
\alias{outlier}
\title{Check presence of outliers using Bonferroni-Holm tests for the adjusted p-values}
\usage{
outlier(resid, alpha = 0.05)
}
\arguments{
\item{resid}{residuals (e.g. lm.object$residuals)}
}
\description{
Check presence of outliers using Bonferroni-Holm tests for the adjusted p-values
}
