% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reliability_index}
\alias{reliability_index}
\title{Reliability index}
\usage{
reliability_index(dat, Ann, pheno)
}
\arguments{
\item{Ann}{Annicchiarico output}

\item{pheno}{character with trait ID}

\item{df}{data.frame}
}
\description{
Reliability index
}
