% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{safety_first}
\alias{safety_first}
\title{Safety-first index}
\usage{
safety_first(df, pheno, design, alpha = 0.95)
}
\arguments{
\item{df}{data.frame}

\item{design}{block or lattice}

\item{alpha}{numeric}

\item{trait}{Random variable}
}
\description{
Safety-first index
}
