% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{selection_gain}
\alias{selection_gain}
\title{Calculates selection gain}
\usage{
selection_gain(pheno, selected_geno, herdability)
}
\arguments{
\item{pheno}{data.frame with first column as the genotypes 
names and following columns the adjusted means for the evaluated phenotype}

\item{selected_geno}{genotypes selected}

\item{herdability}{number}
}
\description{
Calculates selection gain
}
