% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smith_hazel}
\alias{smith_hazel}
\title{Smith Hazel index}
\usage{
smith_hazel(adj.means, cvf, cvg, weights = NULL)
}
\arguments{
\item{adj.means}{data.frame with first column as the genotypes 
names and following columns the adjusted means for each phenotype}

\item{cvf}{phenotypic covariance matrix}

\item{cvg}{genotypic covariance matrix}

\item{weights}{vector with weight of each phenotype}
}
\description{
Is = wYik + wYij
}
